Move `Population` into libOPHD project
Also move `PopulationPool` into the `Population` folder.
