+++ OPHD/MapObjects/Structure.h
@@ -3,9 +3,10 @@
 #include "MapObject.h"
 
 #include "../Common.h"
-#include "../PopulationPool.h"
 #include "../StorableResources.h"
 #include "../UI/StringTable.h"
+
+#include <libOPHD/Population/PopulationPool.h>
 
 #include <NAS2D/Dictionary.h>
