+++ libOPHD/Population/Morale.h
